refactor: split widget data pace logic
